+++ web/consultor-detalhes.xhtml
@@ -2,9 +2,10 @@
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
 <html xmlns="http://www.w3.org/1999/xhtml"
       xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets">
+      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+      xmlns:f="http://xmlns.jcp.org/jsf/core">
     <h:head>
-        <title>#{mainController.consultorVisualizado.nome}</title>
+        <title>#{sessionBean.consultorVisualizado.nome}</title>
     </h:head>
     <h:body>
         <ui:include src="templates/header.xhtml"/>
@@ -12,22 +13,25 @@
             <table>
                 <tr>
                     <td>Nome:</td>
-                    <td>#{mainController.consultorVisualizado.nome}</td>
+                    <td>#{sessionBean.consultorVisualizado.nome}</td>
                 </tr>
                 <tr>
                     <td>Email:</td>
-                    <td>#{mainController.consultorVisualizado.email}</td>
+                    <td>#{sessionBean.consultorVisualizado.email}</td>
                 </tr>
                 <tr>
                     <td>UF:</td>
-                    <td>#{mainController.consultorVisualizado.uf}</td>
+                    <td>#{sessionBean.consultorVisualizado.uf}</td>
                 </tr>
                 <tr>
                     <td>Cidade:</td>
-                    <td>#{mainController.consultorVisualizado.cidade}</td>
-                </tr>
+                    <td>#{sessionBean.consultorVisualizado.cidade}</td>
+                </tr>                
             </table>
         </div>
+        <h:form>
+            <h:commandButton class="btn btn-info add-saldo" value="Contratar" action="#{sessionBean.criarTransacao()}"/>
+        </h:form>
     </h:body>
 </html>
