+++ web/contrato-servico.xhtml
@@ -1,0 +1,20 @@
+<?xml version='1.0' encoding='UTF-8' ?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:h="http://xmlns.jcp.org/jsf/html">
+    <h:head>
+        <title>Contrato Serviço</title>
+    </h:head>
+    <h:body>
+        <h:form>               
+            <h3 style="margin-top: 0">Fale com o Consultor</h3>
+            <p>Deixe seus dados para contato e o que deseja para que possa retornar o contato com uma proposta inicial.</p>
+            <h:inputTextarea rows="10" cols="50"  />
+
+            <h3>Resposta Consultor</h3>
+            <h:inputTextarea rows="10" cols="50" />
+
+        </h:form>
+    </h:body>
+</html>
+
